requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
httpx>=0.24.0

# MongoDB
motor>=3.3.0
pymongo>=4.6.0

# Cloudinary
cloudinary>=1.36.0

# Utilities
python-dateutil>=2.8.2

# For Render deployment
gunicorn>=21.0.0
